% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipv6utils.R
\name{ipv4_to_mapped_ipv6}
\alias{ipv4_to_mapped_ipv6}
\title{Converts IPv4 to mapped IPv6
NAs will be mapped to NAs.}
\usage{
ipv4_to_mapped_ipv6(v4)
}
\arguments{
\item{v4}{Vector of IP v4 Addresses}
}
\value{
List of IP Addresses mapped IP v6 format
}
\description{
Converts IPv4 to mapped IPv6
NAs will be mapped to NAs.
}
\examples{
ipv4_to_mapped_ipv6('127.0.0.1')
ipv4_to_mapped_ipv6(c('127.0.0.1','192.168.1.1'))
}

