% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipv6utils.R
\name{mapped_ipv6_to_ipv4}
\alias{mapped_ipv6_to_ipv4}
\title{Converts mapped IPv6 to IPv4
NAs will be mapped to NAs.}
\usage{
mapped_ipv6_to_ipv4(v6)
}
\arguments{
\item{v6}{Vector of IP Addresses in mapped IPv6 format}
}
\value{
List of IP Addresses in dotted notation IPv4 format
}
\description{
Converts mapped IPv6 to IPv4
NAs will be mapped to NAs.
}
\examples{
mapped_ipv6_to_ipv4('::FFFF:0:127.0.0.1')
mapped_ipv6_to_ipv4(c('::FFFF:0:127.0.0.1','::FFFF:0:192.168.1.1'))
}

