% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipv4utils.R
\name{par_cidr_to_ipv4}
\alias{par_cidr_to_ipv4}
\title{Converts IP v4 ranges in CIDR format to list of IP v4 addresses in dottend notation in parallel}
\usage{
par_cidr_to_ipv4(cidr)
}
\arguments{
\item{cidr}{Vector of IP Ranges in CIDR format}
}
\value{
List of List of IP Addresses in dotted notation format with length same as input
}
\description{
Converts IP v4 ranges in CIDR format to list of IP v4 addresses in dottend notation in parallel
}
\examples{
par_cidr_to_ipv4('127.0.0.1/24')
par_cidr_to_ipv4(c('127.0.0.1/24','192.168.1.1/25'))
}

