% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipv4utils.R
\name{par_ipv4_to_long}
\alias{par_ipv4_to_long}
\title{Converts ipv4 dottend notation to numeric in parallel}
\usage{
par_ipv4_to_long(i)
}
\arguments{
\item{i}{Vector of IP Addresses in numeric format}
}
\value{
List of IP Addresses in dotted notation format with length same as input
}
\description{
Converts ipv4 dottend notation to numeric in parallel
}
\examples{
par_ipv4_to_long('127.0.0.1')
par_ipv4_to_long(c('127.0.0.1','192.168.1.1'))
}

