% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipv4utils.R
\name{par_long_to_ipv4}
\alias{par_long_to_ipv4}
\title{Converts int to ipv4 dottend notation in parallel
NAs will be mapped to NAs.}
\usage{
par_long_to_ipv4(i)
}
\arguments{
\item{i}{Vector of IP Addresses in numeric format}
}
\value{
List of IP Addresses in dotted notation format with length same as input
}
\description{
Converts int to ipv4 dottend notation in parallel
NAs will be mapped to NAs.
}
\examples{
par_long_to_ipv4(2130706433)
par_long_to_ipv4(c(2130706433,2130706434))
}

